% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-grades.R
\docType{data}
\name{grades}
\alias{grades}
\title{fake exam grade data}
\format{A data frame of fake data with 100 rows and 3 variables:
\describe{
  \item{anxiety}{Anxiety rating}
  \item{preparation}{Preparation rating}
  \item{exam}{Exam grade}
}}
\usage{
data(grades)
}
\description{
fake exam grade data
}
\keyword{datasets}
