% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph3D.R
\name{mr.plot}
\alias{mr.plot}
\title{Creates moderated regression 3D graph}
\usage{
mr.plot(
  lm_object,
  criterion,
  predictor1,
  predictor2,
  center.predictors = FALSE,
  axis.labels = NULL,
  cam.position = NULL
)
}
\arguments{
\item{lm_object}{Name of independent variable 1 column in data frame}

\item{criterion}{Name of independent variable 2 column in data frame}

\item{predictor1}{Name of dependent variable column in data frame}

\item{predictor2}{Project data frame name}

\item{center.predictors}{test}

\item{axis.labels}{test}

\item{cam.position}{A list with theta (degrees), phi (degrees), and distance values. Suggest default distance of 3.}
}
\value{
plotly object
}
\description{
Creates moderated regression 3D graph
}
\examples{
head(grades)

lm_object <- lm(exam ~ anxiety + preparation + anxiety*preparation, data = grades)

# Example 1: Not mean centered
graph3D <- mr.plot(lm_object,
              criterion = exam,
              predictor1 = preparation,
              predictor2 = anxiety)
print(graph3D)

# Example 2: Mean centered
graph3Dcentered <- mr.plot(lm_object,
                        criterion = exam,
                        predictor1 = preparation,
                        predictor2 = anxiety
                        center.predictors = TRUE)
print(graph3Dcentered)
}
\references{
Cohen, J., Cohen, P., West, S. G., & Aiken, L. S. (2003). Applied multiple regression. Correlation Analysis for the Behavioral Sciences (3rd Edition). Lawerence Earlbaum Associates: London. ISBN-13: 978-0805822236
}
