% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fast.int}
\alias{fast.int}
\title{Creates a table of means and standard deviations for a 2-way ANOVA design in APA style}
\usage{
fast.int(lm_object, criterion, predictor, moderator, axis.labels = NULL)
}
\arguments{
\item{lm_object}{Name of independent variable 1 column in data frame}

\item{criterion}{Name of independent variable 2 column in data frame}

\item{predictor}{Name of dependent variable column in data frame}

\item{moderator}{Project data frame name}

\item{axis.labels}{test}
}
\value{
APA table object
}
\description{
Creates a table of means and standard deviations for a 2-way ANOVA design in APA style
}
