% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fast.int}
\alias{fast.int}
\title{Creates a MMR plot and output}
\usage{
fast.int(data, criterion, predictor, moderator,
  center.predictors = FALSE, axis.labels = NULL)
}
\arguments{
\item{data}{Name of data frame for interaction}

\item{criterion}{Name of independent variable 2 column in data frame}

\item{predictor}{Name of dependent variable column in data frame}

\item{moderator}{Project data frame name}

\item{center.predictors}{test}

\item{axis.labels}{test}
}
\value{
plotly object
}
\description{
Creates a MMR plot and output
}
\examples{

fast.int(data = grades,
         criterion = exam,
         predictor = preparation,
         moderator = anxiety,
         center.predictors = TRUE)

}
