% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fast.int}
\alias{fast.int}
\title{Creates regression output for overall moderation and simple slopes analysis (2D/3D plots)}
\usage{
fast.int(data, criterion, predictor, moderator,
  center.predictors = FALSE, axis.labels = NULL)
}
\arguments{
\item{data}{Name of data frame}

\item{criterion}{Name of criterion variable}

\item{predictor}{Name of predictor variable}

\item{moderator}{Name of predictor variable}

\item{center.predictors}{Boolean. Indicate if predictors hsould be centered. Default is FALSE.}

\item{axis.labels}{Optional. Override column names as axis labels by sending in a list of new names.}
}
\value{
Returns a list with many objects (tables/graphs) that are displayed on screen when printed.
}
\description{
Creates regression output for overall moderation and simple slopes analysis (2D/3D plots)
}
\examples{

# Compare results to Table 7.4.1 from Cohen, Cohen, West, and Aiken (2003)

fast.int(data = cohen_exercise,
         criterion = endurance,
         predictor = age,
         moderator = exercise,
         center.predictors = TRUE)
# or

new_axis_labels <- list(criterion = "Endurance",
                        predictor = "Age",
                        moderator = "Exercise")

fast.int(data = cohen_exercise,
         criterion = endurance,
         predictor = age,
         moderator = exercise,
         center.predictors = TRUE,
         axis.labels = new_axis_labels)

}
