% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fast.plot}
\alias{fast.plot}
\title{Creates a MMR plot}
\usage{
fast.plot(lm_object, criterion, predictor, moderator,
  center.predictors = FALSE, axis.labels = NULL)
}
\arguments{
\item{lm_object}{Name of independent variable 1 column in data frame}

\item{criterion}{Name of independent variable 2 column in data frame}

\item{predictor}{Name of dependent variable column in data frame}

\item{moderator}{Project data frame name}

\item{center.predictors}{test}

\item{axis.labels}{test}
}
\value{
plotly object
}
\description{
Creates a MMR plot
}
\examples{
head(grades)

lm_object <- lm(exam ~ anxiety + preparation + anxiety*preparation, data = grades)

# Example 1: Not mean centered
fast.plot(lm_object,
         criterion = exam,
         predictor = preparation,
         moderator = anxiety)

# Example 2: Mean centered
fast.plot(lm_object,
         criterion = exam,
         predictor = preparation,
         moderator = anxiety,
         center.predictors = TRUE)

}
